Don't instantiate Request multiple times.
All the responsibility for pulling apart the request has moved into
Request (previously RequestParser), and this now gets instantiated
earlier and passed into the operations. This gets rid of the (untested)
EnvHelper.

There are some pending tests for Request that need filling out to cover
some untested behaviour.

Next I'd like to split ObjectStore so responsibilites for buckets and
objects are handled separately.

See the previous commit for further work that's needed.